Silence paranoid clang cpp warnings for the ∧-character
Recent versions of clang complains about "unicode-homoglyphs" for the ∧ character,
(because it can be confused with `^`), which makes sense when it appears in code, but
not so much in comments. Of course, clang doesn't know about Haskell comments...

Fix: add a single quote on the lines containing ∧

An alternative would be to pass -Wno-unicode-homoglyphs to cpp, but I'm not sure how
well that would play with other preprocessors.