[ fix ] ignore duplicate declarations of libraries
Some simple installers seem to append the path to the stdlib's .agda-lib
file at the end of libraries-VERSION in AGDA_DIR no matter whether it is
already there or not.

This can lead to puzzling errors when re-installing the library: suddenly
there are two equal paths and Agda starts complaining:

> "Ambiguous library 'standard-library'. Could refer to any one of
>  standard-library (/usr/local/lib/agda/standard-library.agda-lib)
>  standard-library (/usr/local/lib/agda/standard-library.agda-lib)"

This PR nubs the list of filepaths. The hardest part was to get the flags
right for the test case reproducing that situation...