[ generalize ] deal with context properly
Generalized types are checked in a context extended with a single new variable,
which is later substituted by a record value containing the generalized variables.
This way we have correct context all the way and don't run into problems with things
(like pattern lambdas and non-generalizable metas) being created in the wrong context.