[ fix #3338 ] defer axiom generation for lone signatures
If a lone signature is in an instance block, it will first be made
`nice` and *then* modified to be marked as an instance argument.
If we pre-generate axioms at the point of nicification of signatures
then we miss out on that extra info.

Solution: revert back to storing the minimum amount of information
in `DataRecOrFun` and use the `NiceDeclaration` the lone signature
has been turned into as a template for our `Axiom`.