- simpler rendering of pattern matching on nullary tuples
`() => exp' is rendered now as '_ => exp'.

- do not emit line breaks in minified code at all

- add 'const' keyword before function declarations in the generated JS code

- Use '0' instead of 'null'

'0' is shorter notation for expressions which are never used

- add --js-output flag to the JS backend

If this flag is turned on, the compiler generates one big JS file
instead of one JS file for each Agda module.

dead code elimination for the JS backend

Dead code elimination works only for concatenated JS output, i.e. if
`agda` is called with `--js-output`.

Dead code elimination will keep the definitions which are used
(transitively) in the definition of external symbols.

- The new flag `--js-externals` tells the JS backend where to find the
file which contains the list of external symbols.

If the list is empty or no `--js-external` are given then all definitons
are kept.

Top level definitions will be renamed to two-character names plus an
optional number:

    aa, ab, ac, ..., ba, bb, bc, ..., aa1, ab1, ac1, ...

Renaminging is done only if external identifiers are provided with
`--js-externals`.

External identifiers are renamed to the given name in the file which
lists external identifiers.

The format of the external identifiers file is:
Each row contains an external identifier and its new name separated by a
space character.

agdaRTS functions can be renamed too.

Example externals file with short renames:

    agdaRTS.primSeq S
    agdaRTS.primIntegerFromString I
    agdaRTS.uprimIntegerPlus IPlus
    agdaRTS.uprimIntegerMinus IMinus
    agdaRTS.uprimIntegerMultiply IMul
    jAgda.Prelude.Bool.true T
    jAgda.Prelude.Bool.false F