[ fix #2044 ] improve error messages for failed instance search
Distinguish between candidates that fail because they have the wrong
type and candidates that fail due to some constraints. If there's a single
one of the latter candidates (and on valid instance is found) report the
error from the constraint. If there are multiple of the latter report the
candidates together with the corresponding error.