[ Setup.hs ] type-check Primitive in a build phase
Instead of type-checking after registration, we type-check Builtin and
Primitive modules in the build phase.  The newly created `.agdai` files are
then copied in copy/install phases to their final location.

This is approach also used by `happy` package. I tested it with
GHC-8.4.4 with:

- stack build
- stack install
- cabal v2-build
- cabal v2-install
- cabal sandbox init && cabal install

cabal-install version 2.4.1.0
stack Version 1.9.1

The only downside is development tree is polluted with `.agdai` files
too. As they are already in `.gitignore` I don't think it's a big issue.