Fixed #3648.
I did not include a test case. I do not want to include a broken
.agda-lib file in the root directory of the Agda source tree, and I do
not want to include an extra build of Agda in the test suite. It might
be possible to only test this part of the build, but I do not want to
add a potentially brittle step to the test suite (brittle because
people build and install Agda in different ways) just to guard against
this minor problem.