Relax ghc-8.2.0 check that can apply to all >=8.*
The `Any` type doesn't show up in Hackage for `base-4.9.0.0` like it
does for `base-4.11.0.0`, but it _is_ exported since all of `module
GHC.Prim` is exported by `GHC.Exts`.
  * https://hackage.haskell.org/package/base-4.11.0.0/docs/GHC-Exts.html#t:Any
  * https://hackage.haskell.org/package/base-4.9.0.0/docs/src/GHC.Exts.html#line-30

Relaxing this check helps (in future commits) to reduce the number of
build specializations relying on the C preprocessor.

Note that there was no specific 8.0.0 release, but it serves as a
convenient version boundary.