Use Semigroup/Prelude.<> vs PrettyPrint.<>
Using Prelude's `<>` on pretty printing is defined as the same function
as `PrettyPrint.<>`
(https://hackage.haskell.org/package/pretty-1.1.3.6/docs/src/Text.PrettyPrint.HughesPJ.html#line-133)

with the important difference that it is left associative (infixl)
instead of the Prelude's right associative (infixr) definition.
See: https://www.haskell.org/pipermail/libraries/2011-November/017066.html

This becomes slightly awkward when combining with `<+>`, but seems to be
a worthwhile trade-off. (Alternatively, we could redefine `<+>` to be
infixr also to remove the ambiguity; but that seemed to be slightly more
heavy-handed for a first pass).

Note that some elements of this are redundant once <GHC-8.4 support is
dropped. The `import Monoid hiding ((<>))` can be unhidden, and
re-exporting `Semigroup((<>))` is unnecessary in 8.4 and above.