[ghci] Auto-load version-specific .ghci files
This allows one to run `ghci` using the same flags as in the Cabal file
for version-flagged GHC versions. It also allows us to turn on more
recent features for developers without breaking backwards compatibility.

The `.ghci` file is loaded first, it checks the GHC version, and loads
the latest matching `.ghciXXX` file. Currently each `.ghciXXX` file
simply calls the version below it until it reaches the
lowest-common-denominator.