Fix clashing symbols when loading in interpreted ghci
Several modules rely on `.hs-boot` files' limited sets of imports to
avoid clashing smbols: when compiling, if you "import {-# SOURCE #-}"
you expose only the symbols in `.hs-boot`, versus if you normally
import.

This is weird but normally fine. However, with GHC (8.6.3, anyway), if
you load files via `:load` in GHCi with `:set +c` (collecting type
info), then all of the symbols become visible, leading to several
"Ambiguous reference" errors.

This change is intended to be a total no-op, except that it allows the
codebase to play nicely with GHCi.