[travis/stack] Run cpphs/debug as separate job
Previously the jobs would run serially with a `stack clean` in between.
Hopefully this will help the builds to run faster.

Note that this now only runs that test for the latest GHC, not for each
version. Improving CI turnaround should be a good tradeoff, though,
since `cpphs` bugs are unlikely to be *super* version specific. If they
are, we could add additional permutations (running in parallel, not
sequentially!)