[ fixed #3818 ] by passing A.ModuleName to openModule (where known)
This heals the previous information loss in the treatment of

  open import M

where `open` did not receive the information that `M` is an external
module, and thus could not distinguish it from an internal module with
the same name in scope.

By note forgetting the resolution of M as abstract module name, this bug
is automatically fixed.