[ fix #3829 ] case-split aware of later clauses too (#3834)
The new strategy makes a difference when the user intentionally introduces a
pattern causing later ones to become unreachable and decides to split this
new pattern.

Instead of getting all possible cases, they only get the ones which are not
already covered by the unreachable clauses. This makes it easier to extend a
datatype with a new constructor and to add extra clauses to existing definitions
anywhere but at the end of their existing list of clauses.