[ comments, fixed #3861 ] clarified question about semantics of Flexible [ci skip]
Andreas: A flexible variable occurrence is decorated by a set of meta
variables.  This was originally introduced by @jespercockx in
f305ee9d05b26dd5e4f74a12927e0eabd4988a8d to use in REWRITE pattern
matching.  What is its intended semantics?

The `FlexRig` type is (supposed to be) a semiring with addition
combining information from different subterms and multiplication
combining the context of an occurrence with the occurrence.  Currently,
multiplication `composeFlexRig` is explicitly implemented using the
union of the metasets.

For example `x` occurs `Flexible [X,Y]` in `X (Y x)` (where capitals
denote meta vars).  Spelled out, `x` occurs `Flexible ms` if there is a
chain of meta-applications given by metas `ms` from `x` to the current
point of view.  Is this the intended semantics ("and")?

Addition: How should `x` occur e.g. in `(X x , Y x)` ("or")?

The _free_ semantics would use a semiring instead of a meta set.
However, what we actually need here depends on the intended use.

UPDATE: My commit 031a490c05cee4ec6153975b5931fec66efbbec6 contains a
tentative replacement for `Ord.max` called `addFlexRig`, also using
union of meta sets.

Jesper: Well the comment of the constructor says In arguments of
metas. The set is just there to keep track in the arguments of which
metas the variable occurs. This allows us to retry the free variable
check when one of these metas is instantiated, hopefully making the
variable no longer free.

At least that's the theory, in practice most of this information is
currently forgotten as there is only room for a single metavariable in a
Blocked tag.

In theory, the MetaSet could be replaced by a richer (free?) semiring
structure which allows us to keep track of precisely when we should
retry the check (e.g. we should retry if m1 AND m2 are instantiated, OR
if m3 is instantiated). But this would currently be a complete waste
because of the point above.

Andreas:  So then x should also occur Flexible [X, Y] in (X x, Y x).
Addition and multiplication is the same operation onMetaSet. The
semantics is that the status of variable x can change when one of the
metas gets instantiated.