[ fix #3866 ] Do not mess with subscripts
Rather than trying to convert back and forth between unicode subscripts
and indices, we change the notion of `nextSuffix`. When --no-unicode is
turned on, `nextSuffix` generates indices rather than subscripts when
the previous suffix was empty.

If the --no-unicode user inputs some subscripts then we respect their
choice: garbage in, garbage out!