[ fixed #3872 ] {-# OPTIONS --no-forcing #-} is now legal
Since `--no-forcing` only steers the analysis which constructor
arguments should be forced, but not what the compiler does with it,
there is no risk of incompatibility of interfaces if `--no-forcing`
modules use `--forcing` modules or vice versa.  In fact, forcing could
be disabled (or enabled) on a per-definition basis, similar to
`{-# NO_POSITIVITY_CHECK #-}` and friends.
A first step is to enable it on a per-file basis.