agda-input.el: Add some missing characters
Some characters were missing because their numeric Unicode value is out of
sequence of the others in that set, because the character might be doing
double-duty.

This also adds comments for some characters that are just plain missing
from Unicode, as of the current standard (12.1), so that nobody else
goes on a wild goose chase looking for them.