[ new ] more liberal syntax in telescopes
* Extending the switch-to-hidden syntax in typed bindings to
  instance arguments i.e. (a {b c} d : Nat) used to be valid,
  now (a {{b c}} d : Nat) also is.

* Allowing as patterns in binding positions in telescopes.

NB: this is actually a language change, cf. the type of `j` in
the test case:

j : ∀ a (e , f) → Nat

used to mean (here `,` is an identifier like any other)

j : ∀ a e , f → Nat

but now it means:

j : ∀ a _@(e , f) → Nat

I have run the new parser on the whole test suite (including the
stdlib) and it has not broken anything. Was anyone actually using
the `∀ a (b c) d e f` syntax before?

The next step is to actually make the as-patterns do something.

[ci skip]