[ #3792 ] refactored conErased to Maybe [Bool]
It is now Nothing before the treeless erasure analysis has run on the
constructor.  Currently we are lenient towards accesses of Nothing: we
return the information that no argument is erased.  This is because the
erasure analysis is not run on all constructors, for instance not on the
list constructors [] and (:) in Agda.Builtin.List.

In the future, the invariant that conErased should be set for all
constructors to Just... at some point before used in compiled code could
perhaps be established.