[ #3792 ] fixed faulty MAlonzo compilation of constructor applications
The previous code for `term` compilation relied on the assumption that
TCon are never under a TCoerce.  This assumption was false.

The new code for `term` has a perspicuous structure, using a new
coerceAppView.  Unnecessary eta-expansions of constructors are no longer
produced.

Bugs like the present can be avoided in the first place by implementing
robust and well-structured syntax traversals, and VERIFYING invariants
one is relying on by making ample use of assertions.

__IMPOSSIBLE__ is your friend!