[ fix #1097 ] more liberal syntax in telescopes (#3884)
ping #2361

* [ new ] more liberal syntax in telescopes

* Extending the switch-to-hidden syntax in typed bindings to
  instance arguments i.e. (a {b c} d : Nat) used to be valid,
  now (a {{b c}} d : Nat) also is.

* Allowing as patterns in binding positions in telescopes.

NB: this is actually a language change, cf. the type of `j` in
the test case:

j : ∀ a (e , f) → Nat

used to mean (here `,` is an identifier like any other)

j : ∀ a e , f → Nat

but now it means:

j : ∀ a _@(e , f) → Nat

I have run the new parser on the whole test suite (including the
stdlib) and it has not broken anything. Was anyone actually using
the `∀ a (b c) d e f` syntax before?

The next step is to actually make the as-patterns do something.

[ci skip]

* [ broken ] Bubbling up the patterns in concrete syntax

Not sure whether to elaborate them away now using let-bindings
or to propagate them to Abstract syntax too.
[ci skip]

* [ broken ] Concrete to abstract

* [ broken ] Compiling again

And I even think I may have gotten the TypedBindings translation
right. However I currently get an internal error in ConcreteToAbstract.

[ci skip]

* [ partial fix ] Typed bindings are now working

Do not forget to bind the variables bound in the pattern!

* [ partial fix ] support for patterns in lambdas

* [ refactor ] typeError . GenericError -> genericError

* [ fix ] pattern-matching binders in teles & lambdas

* [ fix ] raise error for pattern in data/record telescopes

* [ highlightin ] binders (or at least trying to)

* [ test ] new examples

* [ test ] shadowing

* [ test ] stronger shadowing test

* [ fix ] type annotation using 'Binder'

* [ fix ] merge errors

* [ fix ] use `bindToConcrete` to get right renaming behaviour

* [ test ] more complex shadowing example

[ci skip]

* [ changelog ] documenting the new feature

[ci skip]

* [ manual ] entry for irrefutable patterns in telescopes

[ci skip]