[ fix #2610 ] Build directory for version specific agdai files
If an .agda file is part of a project with an identifiable root
(i.e. if there is an agda-lib file in any of the directories above it)
then we create a _build/VERSION directory at that root and we
write the .agdai file storing its interface in it.

If it does not have an identifiable root then we simply store the
.agdai interface file alongside the file itself.

Additional refactoring: disambiguating between AbsolutePath standing
for source files and the ones standing for interface files.