[ fixed #4159 ] by uncurrying metas in addConstraint
Adding constraints will now simplify constraints in assignment form
where the meta is applied to some projected variables.  E.g.

  r : record { f : A } |- X (r .f) = rhs

will be replaced by

  "(r .f)" : A |- X "(r .f)" = [ record { f = "(r .f)" } / r ] rhs

Here "(r .f)" is a name for a new variable created for the field f of r.

This mechanism (expandProjectedVars) was already at work in
MetaVars.assign, but not available for size constraints.  In particular,
this mechanism is needed for the generalization feature which uses a
variable of record type to store all the generalized variables.

As a consequence of this change, constraints displayed to the user will
sometimes have the funny generated variable names which try to mimic the
term they represent.

At the moment, there is still a doubtful implementation of the
substitution, inherited from the current implementation of this
expandProjectedVars feature:  The new constraint lives in the context

  inTopContext $ addContextTel tel

where tel represents the new context.  This should be replaced by

  updateContext <subst> (const tel)

to be safe with checkpoints and parameter substitution.  However, the
API for contexts does not accept a telescope here, and there is no
function to turn a telescope into a context.

Either such a function has to be provided (but note that the CxtIds are
then not preserved!) or the expandProjectedVars mechanism has to be
changed to work with context instead of telescopes.