[ #4189 ] bring back qualified record constructors
But leave the OnlyQualified hack in Hades.

We now exploit the "open public" feature of modules that let's us have
names in modules that ignore the record parameters.

Technically, we simply link the concrete constructor name inside the
record module to its definition in the parent module.

Maybe this is how the feature should have been implemented in the first place.