[ fixed #4267 ] record expressions need the record type "in scope"
This fixes #4267; but the root cause of #4267 is the unhygienic import
handling of Agda, see #4272.

The fix filter possible candidates for (record) types of a record
expression by Agda.Syntax.Scope.Base.isNameInScope.

It is not entirely clear what this check entails (this function is
entirely undocumented); for instance, it does not entail that the name
that Agda infers for the record type could actually be written by the
user (see test case Issue4267a).

However, it seems to do the job to fix the present bug (test case
Issue4267 passes now without unsolved metas).