[ fix #4390 ] [ fix #4401 ] Perform subtyping check on context of meta before solving
When solving an equation `_X y1 .. yn =?= v` with
`_X := \y1 .. yn -> v`, we need to check that the types of y1 .. yn
are actually equal to the context where _X was created, for otherwise
v might not be well-typed in the context of the meta.

An alternative (more permissive) solution would be to re-check in the
context of the metavariable using checkInternal, but this might be too
expensive in general.