[ fixed #4451 ] refactor: merge results of lookups instead of merging maps
Given that the imported signature can be large, it seems more efficient
to not subject it to merging with a smaller local hashmap.  Rather, we
reconcile the results of looking up the QName in the 3 differend
hashmaps involved here.

The law underlying this refactoring is:

    unionsMaybeWith f . map (lookup k)     {new}
      = lookup k . unionsWith f            {old}

This refactoring seems to at least halven the time spent on highlighting
the standard library!