[ benchmark ] Haskell scripts to create import benchmarks
A few short Haskell scripts to create benchmarks that accumulate
imports, in the sense that file #n imports all the previous ones.

Agda spends 15mins when driving the number to 1000, with the bulk of
time on serialization.

Haskell copes even worse with such massive imports, I did not await GHC
finish for n=1000.

Some kind of quadratic behavior is of course involved here, and could
likely dominate the overall performance.

These benchmarks are not hooked up to the test suite as of now.
[ci skip]