[ fix #3417 fix #4348 ] highlight warnings as they are created
Previously, *all* warnings of a file were highlighted upon successfully
type-checking a declaration.  Problem was: if the declaration following a
termination error had a hard error, the termination problems were not
highlighted.

Now, we simply highlight warnings as they are created using the
"warning" method.  This fixes the highlighting repetition for warnings
observed in #4348, and also the missing warning highlighting upon hard
errors.