[ fixed #4518 ] More information from operator pattern parser
Give more information why "Could not parse the lhs".
If the LHS contains a sub-pattern that failed to be a valid pattern, we
report this pattern to the user.

Technically, I replaced `Maybe` used as error monad by `Either Pattern`
where this pattern is the first offending one.

Sometimes this pattern is the whole LHS, then it should be omitted.
Thus, there is room for improvement.  We could switch to
`Either (Maybe Pattern)`.