[ #4569 ] MWE for failure of Mac OS X install with dynamic linking
In directory

  test/Bugs/Issue4569

type

  make

which will try to install the HelloWorld program with a dynamically
linked executable hello-world.

This fails on Mac OS X since the @rpath to locate the HelloWorld library
is set to be in the installation location (.cabal/lib/...) but when the
executable is run by the Setup.hs script the library has not been copied
there yet.  It would need an additional @rpath entry that points to the
build location (dist/build/).

Use

    otool -l <executable>

to get information where dynamic libraries are searched.

[ci skip]