[ refactor for #2038 ] highlighting: nameInfo/theRest w/o geniplate
This refactoring implements the extraction of highlighting information
from abstract syntax by a manual traversal.  Previously, this traversal
was split into 13 individual traversals automatically generated via
geniplate.

The goal is to get rid of geniplate which has been abandoned by the
original author, putting the burden of maintenance on us.  Getting rid
of geniplate will also dispose of the use of TemplateHaskell for
Abstract.hs, which noticeably slows compilation.

This patch does not reach the goal completely, as the NameKinds map is
still constructed using geniplate.  The NameKind information could
instead be maintained in the ScopeInfo.

A more radical refactoring would get entirely rid of a separate
highlighting pass over abstract syntax and instead record information
about highlighting in the TCState, incremently adding information
whenever the kind of a name is resolved in the scope checker.  We
already do a similar thing for the disambiguation of overloaded symbols
in the type checker.