[ #3821 ] Work in progress.
An apply primitive has been implemented, but it has not been
documented in the changelog or the user manual.

Quick testing suggests that this apply primitive might be slower than
an implementation that uses the rest of the reflection machinery. The
other implementation does not automatically fill in missing implicit
arguments, but I suspect that this does not make much of a difference.