[ refactor for #4560 ] forbid split on codata right away
The old noPatternMatchingOnCodata was a separate pass after splitting
was done.  We can more eagerly disallow splits on coinductive
constructors.  We do so for other types of disallowed splits already.

This refactoring saves an extra traversal on the internal patterns.

In connection with this refactoring, I split Syntax.Common.DataOrRecord
into its different uses:

1. Syntax.Scope.Base.DataOrRecordModule
2. TypeChecking.DataTypes.DataOrRecord
3. TypeChecking.Rules.LHS.DataOrRecord with extra information