[ fixed #4560 ] pattern attribute no-eta-equality inductive records
Having both pattern and copattern matching for records without
eta-equality breaks subject reduction.

If you want to pattern match on the constructor for a no-eta record, you
have to declare it 'pattern'.  The default is copattern matching.

The attribute 'pattern' is ignored for 'coinductive' or 'eta-equality'
records.  Recursive records might be declared as no-eta by the
positivity checker; thus, we need to remember the attribute 'pattern'
until the positivity checker has run.  It is stored in the new field

  recPatternMatching :: PatternOrCopattern

of the Record alternative of TCM.Defn.  Ultimately, the choice of
PatternMatching or CopatternMatching is integrated in the field

  recEtaEquality' :: HasEta

This change to Agda affects a dozen test cases and also the standard
library at one place.  The standard library has been patched to be
compatible with this commit.