[ fix ] differentiate untyped & typed patterns
In some cases Agda thinks it's really clever by changing an isolated
`FunClause` of the shape `f = ...` into a `FunSig` (of type `_`) with
a `FunDef`. However in our case it breaks the distinction between the
untyped and typed pattern synonyms.

This fixes this problem by making the type in `FunSig` optional and
only turning it into an underscore in the cases where we are not inside
of a pattern block.