Spend some effort printing typeError and debugPrint messages in TC
Previously all parts were `fsep`d together. Now we split the error up
into lines and hcat the parts of a line and vcat the result. A side effect
of this is that there is no longer a space inserted before and after terms
and names, so some error messages might need to insert extra spaces
to not look weird. On the other hand you can now construct error messages
that you couldn't before.