[ #4547 refactor ] Use List2 in RawApp[P]
An application naturally is a list of at least 2 elements.

No more removeSingletonRawAppP!  Eliminated ~10 corner cases.

Change to Concrete.Definitions:

- Since EllipsisP is no longer wrapped in a RawAppP which stores its
range even after replacement, we need to set the range of the ellipsis
to the pattern we substitute for the ellipsis.

- As a side effect, the ellipsis ... maybe be highlighted as function now,
allowing jump-to-definition.  Can be considered a feature rather than a
bug.