[ #4736 ] Handle Ctrl-c in test runner gracefully
By default, all spawned Agda processes are part of the same process group
as the test runner. If one presses Ctrl-C, the terminal will send
SIGINT to all process in that process group. It can happen that this
kills the Agda process before the test runner, in which case the test
runner will try to update the golden file with the new exit code.

To avoid this, we spawn the Agda processes in a new process group.
On Ctrl-C, the test runner still receives the SIGINT and the
process(-extras) library ensures that the spawned processes
are killed in that case.