[ #4135 ] LHS constructor disambiguation: solve metas only uniquely
The regression introduced in 2.5.4 (likely
fea1dd1f1889fb122320314c0bb7d6a6126f6dd5) used 'noConstraints' as if
this would prevent any kind of information gain on metas.  However, it
does so only in combination with 'dontAssignMetas'.

The present test case illustrates the problem.  The constructor can be
disambiguated if either '?0 := true' or '?0 := false'.  The previous
algorithm would happily first set '?0 := true' since this did not leave
any new unsolved constraints, and then fail on the second constructors
since now the parameter was "known" to be 'true' and could thus not be
'false'.  Thus, the algorithm would believe it in unique disambiguation.
Alas, beliefs can be wrong if derived via an unsound methodology.

The new algorithm tries disambiguation first with properly disallowing
information gain (dontAssignMetas . noConstraints).  If this does not
work, it allows information gain but packages this (TCState) up with the
disambiguation candidate.  Only if just one candidate remains, the
information gain takes effect.

Duplicate disambiguation candidates are filtered out not only if their
original constructor name is the same, but also if their type (with
parameters is the same).  When comparing two types, we can use the
information gain (new TCState) if at most one type comes with it.  If
both disambiguations needed information gain we consider them
different (even if the gain was the same, which we cannot efficiently
check, or do not consider worth at this point).

This patch fixes the constructor part of #4135.