git: Sort .gitignore with sort -u
The comment says to keep it sorted, but it wasn't, and was unclear what
sorting was prefered.

This sorts the .gitignore file consistently and adds an additional
comment specifying the command to do so (using `sort -uf`, which means
"unique and ignore case", and works the same for GNU/BSD/macOS).