make: Fix incorrectly quoted loop in test script
Several quoting issues in this one line:
  for file in ls `$*.*agda`; do touch $file; done ;

Would expand in make to the following shell script (replacing $* with
the make target, and $f replaced an undefined/empty variable):
  for file in ls `Foo.*agda`; do touch ile; done ;

Which bash interprets as: for each word in this list, consisting of the
word "ls", and (whatever comes out of executing a command named
"Foo.*agda"), touch the file named "ile".

This has been updated to just touch any files matching that glob, but
not attempting to create a file if none match.