make: Load deprecated mk/parallel-test.mk if exists
The Agda makefiles prior to PR #4841 would load `mk/parallel-test.mk`,
which, if existed, would be expected to define the single variable
`PARALLEL_TESTS`. They did not load `mk/config.mk` or expect that to
exist.

It's simpler to maintain a single source of truth for user
configuration, but we don't want to spring changes on developers who had
customized that option.

So: if the new `mk/config.mk` file does not exist, but the old
`mk/parallel-tests.mk` does, load it and gently nudge them toward
putting creating a `mk/config.mk`.

We don't create it automatically because this loading phase may not be
atomic and we don't want to clobber things that aren't build targets.

This commit should be reverted at some point in the mid future, leaving
`mk/config.mk` as the single source of truth.