refact: Refer to unitModality/etc. instead of mempty
Because Modality/Quanitity/Relevance/Cohesion form semirings, the monoid
referred to by `mempty` is ambiguous. It also differs from
`defaultModality`, though `mempty` is traditionally used in that way.
This weens away from the dependency on that choice and makes it explicit
whether it's the additive or multiplicative unit being used.

Note that some cases may have _meant_ to use `defaultModality` instead
of `mempty` -> `unitModality`. However this intentionally avoids
altering behaviour.