stack: Include src/fix-whitespace in extra-deps
This fixes an error in which `make` failed with stack due to trying to
build `stack build fix-whitespace`, which it had no knowledge of.

It's added to the `extra-deps` so that it does not need to be built by
default. (It's only used by `make fix-whitespace` which will correctly
attempt to build/install it if necessary).