travis: Handle argument list expansion more safely
Bash has a lot of subtle text expansion rules that were not being taken
into account. For the existing set of arguments that was fine, but would
completely fall apart in the presence of an argument containing spaces,
like `--ghc-opts '+RTS … -RTS'`.

Here those argument lists are stored in bash arrays, and then expanded
again as quoted arrays to avoid additional word-splitting.

[github skip]