[ levels ] remove LevelAtom type
This type was added in an effort to save some computation time by remembering
the status of reduced levels. Given that reducing levels is hardly a bottleneck, and
that this extra layer adds a lot of code, it seems best to remove this experiment.

Benchmarking the standard library doesn't show any statistically significant difference
before and after this patch. If anything it makes things slightly faster.