Replace "file and line" with CallStack for errors
This replaces use of the `AgdaSourceErrorLocation` or separate `file`
and `line` arguments with the GHC `CallStack` provided by
`HasCallStack`.

This is the *same mechanism* the previous `file` and `line` were derived
from, so should have ~the same performance characteristics as before. It
does not require -oprof or similar.

The `CallStack` type provides a little more information than we had
before, function name, column, and module, which are now included in the
error message.

Additionally, if there are several nested `HasCallStack`-constrained
calls, we print as much of that stack as we can, which increases the
utility of that constraint. Note that typically this is still only a
single-level stack, but can now be usefully deeper which helps with
tracing.