[ fixed #4952 ] do not include binding site when highlighting renaming to
Currently, the abstract syntax for renaming clauses like

   open import M renaming (Nat to N)

looks in abstract syntax just

   renaming (M.Nat to M.Nat)

The binding site attached to the `to` part of the `renaming` clause is
identical to the one of the `from` part.  From the one of the `to` part,
the HTML backend produces garbage.

This patch removes that garbage by not including the binding site when
generating highlighting info for the renaming clause.