main: Improve structure of frontend/backend mode validation
This consolidates a lot of the initial "main cli entry" options checking
and structures it in a more-or-less type-safe way.

This also ensures that the "initial invoked file" options is validated
for compatibility early, so that there are fewer forced unwrappings.
(Before, this was impossible because that argument was used as state).

This makes it so that *initial* high-level execution-mode options like
"show help", "show version", or high-level mode selection are all
validated in (roughly) one place. It's not an ideal structure yet: those
flags shouldn't even be _visible_ to most of the other code, to avoid
the chance of some internal behaviour relying on them. One step at a
time.