compiler/API CHANGE: Provide "Maybe FilePath" for interface
For any backend that advertises "only-scope-checking", the interface
file will not have been written. This would trigger an `__IMPOSSIBLE__`
error before even invoking the backend.

The use of this file path for staleness checking is questionable anyway,
but this at least makes it *possible* for scope-only backends.

Note that the JS/GHC backends have been updated, despite not currently
supporting only-scope-checking. (They will just not be invoked in that
case).