compiler+imports: Include source info in CheckResult
This allows another `findFile` to be skipped. Although normally that
would be cached, the cache is not necessarily maintained during
only-scope-checking. And in any case, it seems safer to maintain the
link between "this type-checked interface corresponds to this source" as
much as possible.

Although the `Interface` type keeps the source text around, it does not
keep the path; so things that need (or at least want) the path were
doing redundant lookups after the fact. There are probably many cases of
this.