latex: Gracefully resolve source path
Because the original source path is an optional field used only for the
parser position fields, there's no need to forcefully look it up.

The main reason to prefer *not* looking it up is that it reduces the
dependency on IO and all the other dependencies of `findFile` (like
updating TCM cache state and so on).