ghc: Aggregate visited definitions in GHCDefinition
This is as opposed to revisiting `curDefs`, which is evil external state
which could conceivably mismatch the actual visited definitions. (For
example, if definitions were erased or if definitions might be
introduced by some mechanism like macros in the future).

There's no bug that I know of exhibiting this, but it's an easy scenario
to avoid.