imports/refact: Move "MainInterface" logic out of tcm.warning
Instead of passing the `MainInterface` type to `applyFlagsToTCWarnings`,
which it just uses to determine whether to also ignore
`unsolvedWarnings`, we just pass a (possibly-empty) set through, and
determine that value the only place that MainInterface was passed
through.

This removes a circular import and allows the `MainInterface` type to be
entirely internal to `Agda.Interaction.Imports`.