imports/refact: Factor out "MaybeWarnings" contruction
MaybeWarnings is a weird type that should be replaced by `[TCWarning]`,
or at most a `newtype` wrapping that.

The only difference between it and a list is that the type *appears* to
permit the additional value of `SomeWarnings []`. However the two places
that construct `MaybeWarnings` took pains to convert `SomeWarnings []`
to `NoWarnings`, and nothing ever seemed sure whether to expect
`SomeWarnings []` or not.

This is brittle and confusing. This commit centralizes that logic to one
place, just for clarity.

MaybeWarnings could be replaced with trivial patterns:

```
type MaybeWarnings = [TCWarning]

pattern NoWarnings :: MaybeWarnings
pattern NoWarnings = []
pattern SomeWarnings :: [TCWarning] -> MaybeWarnings
pattern SomeWarnings ws      <- ws@(_:_)
  where SomeWarnings ws@(_:_) = ws
```

However due to cyclic module dependencies forcing the use of the
.hs-boot, that can't easily be done without moving to another module.
(due to https://gitlab.haskell.org/ghc/ghc/-/issues/13322)