imports+warn/refact: Move special-casing of hole warnings
This change is intended to be an exact behavioural refactor, though it
takes some squinting to see why that's the case.

Motivation: removing the `MaybeWarnings` type which is redundant and
used inconsistently with regards to equivalence between `NoWarnings` and
`SomeWarnings []`

Critically, note that everything that _constructs_ `MaybeWarnings`
coalesces those two possible values, and nothing ever filters that list
(anymore: `applyFlagsToMaybeWarnings` used to, but has been removed). So
here we're interested in ensuring they are treated the same when
destructured, and that any filtering of that list preserves those
behaviours.

In this change, `tcWarningsToError` is updated to filter for and detects
unsolved hole warnings, and treats an empty list to mean "no warnings".

That means this part of the diff:
  > ```diff
  > -  case wt of
  > -    SomeWarnings w  -> tcWarningsToError (filter (notIM . tcWarning) w)
  > -    NoWarnings      -> return i
  > -   -- filter out unsolved interaction points for imported module so
  > -   -- that we get the right error message (see test case Fail/Issue1296)
  > -   where notIM UnsolvedInteractionMetas{} = False
  > -         notIM _                          = True
  > -
  > +  tcWarningsToError $ case wt of
  > +      NoWarnings -> []
  > +      SomeWarnings ws -> ws
  > +  return i
  > ```

The original `SomeWarnings w` would always be non-empty. If they were
all holes, it would throw `SolvedButOpenHoles`; if there were non-holes
it would throw `NonFatalErrors`. `SomeWarnings []` would never happen,
so we can treat it the same as "no warnings". So behaviour here is
preserved.