imports: Track + distinguish scope-checked visited interfaces
This adds a new flag to TCM.Base.ModuleInfo, which tracks visited modules,
to distinguish the mode that the module was "visited" with. Using this
flag, we can now safely store scope-checked interfaces.

Previously, scope-checked interfaces were omitted in order (presumably)
to avoid clobbering type-checking results. However, because of this, it
made it impossible for scope-only backends to access any information
about the scope-checked interfaces after checking them! (Aside from the
"main" one).