imports/refact: Only set stCurrentModule in getMainInterface
Rather than setting it on each loaded import.

The `stCurrentModule` state is *only* used by the top-level interactors
(compiler and interaction BasicOps) for determining the main module
name, not any internal state. It should eventually be removed entirely
since those things get the module name in the resulting interface…

In any case: it doesn't need to be set for each module. It's used to
designate the main module, which use relied on it being the last module
to be loaded, which further relied on that module being forcibly
re-checked. Just a whole lot of brittleness.