imports/refact: Separate out getStoredInterface and post-validation
This avoids needing to re-check the disk and redundant cache lookups
work when the real goal is mainly to perform the options and hash
validation.

The log message change in the tests is just due to skipping that
search; the interface is still loaded and validated as usual.

Previously, after type-checking a non-main module, the interface would
be immediately stored so that it could be validated with
`getStoredInterface`. Now that validation is split out from loading, we
don't need to store it in advance specially to allow the loader to find
it.

Note that nothing else will try to load it during validation: otherwise
the externally-loaded interfaces wouldn't work, and/or we would have
import cycles.