imports/refact: Internally use (Interface, [TCWarning]) -> ModuleInfo
Previously all of the internal results were represented as a tuple of
`(Interface, [TCWarning])`. This replaces those with `ModuleInfo`, which
is the type that gets stored in stVisitedModules.

This simplifies some future changes that introduce more stored fields.
As a bonus, we now also store whether the module is primitive or
not, rather than the previous post-hoc mutation.